Update README.md
India's science museums not only celebrate the nation's scientific and technological achievements but also inspire and ignite young minds. These museums act as bridges, connecting the past with the present, traditional knowledge with cutting-edge advancements, and curious learners with the vast world of science and technology. So, whether you are a student, a teacher, or just a curious soul, a visit to these museums promises a journey of knowledge and wonder.